Changed preprocessing script: added pose estimation, removed sign language dataset loading (it has been deprecated)
