+++ scripts/running/run_basic.sh
@@ -10,7 +10,7 @@ scripts=$base/scripts
 dry_run="true"
 
 # set to your desired estimator
-estimator="mmposewholebody" # options: mmposewholebody, mediapipe
+estimator="mediapipe" # options: mmposewholebody, mediapipe
 
 model_name="phoenix"
